+++ NEWS.md
@@ -10,6 +10,7 @@
 - Added error in `_plot_horizon_on_plane` when the horizon cannot be plotted
 - Updated NumPy types (np.int -> int, np.float -> float)
 - Windowing uneven signals is no longer allowed
+- Correctly remove ghost zones for 1-2D HDF5 files
 
 #### New examples